use f64 in evm_set_block_timestamp_interval

+++ crates/anvil/core/src/eth/mod.rs
@@ -581,7 +581,7 @@ pub enum EthRequest {
         feature = "serde",
         serde(rename = "anvil_setBlockTimestampInterval", with = "sequence")
     )]
-    EvmSetBlockTimeStampInterval(u64),
+    EvmSetBlockTimeStampInterval(f64),
 
     /// Removes a `anvil_setBlockTimestampInterval` if it exists
     #[cfg_attr(
